requirements.txt
streamlit
pandas
plotly
pymysql
sqlalchemy
cryptography
python-dotenv
xlsxwriter
openpyxl
bcrypt